requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
psutil>=5.9.0
aiohttp>=3.9.0
